blog-generator/pyproject.toml
[tool.poetry]
name = "obsidian-generator"
version = "0.1.0"
description = ""
authors = ["Milan Suk"]
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = "^3.10"
marko = "^1.2.1"
python-frontmatter = "^1.0.0"
pydantic = "^1.9.1"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
generate = 'src.cli:generate'
